fix(marriage-conditions): cannot marry yourself :(
