+++ libs/application/templates/marriage-conditions/src/fields/InfoAlert/index.tsx
@@ -1,1 +1,17 @@
-export { InfoAlert } from './InfoAlert'
+import React from 'react'
+import { AlertMessage, Box } from '@island.is/island-ui/core'
+import { useLocale } from '@island.is/localization'
+import { m } from '../../lib/messages'
+
+export const InfoAlert = () => {
+  const { formatMessage } = useLocale()
+
+  return (
+    <Box marginY={2}>
+      <AlertMessage
+        type="info"
+        message={formatMessage(m.informationAlertMessage)}
+      />
+    </Box>
+  )
+}
