--- libs/application/templates/marriage-conditions/src/fields/NationalIdWithName/index.tsx
@@ -1,1 +1,0 @@
-export { NationalIdWithName } from './NationalIdWithName'
