+++ libs/application/templates/marriage-conditions/src/fields/index.ts
@@ -3,4 +3,4 @@ export { InfoAlert } from './InfoAlert'
 export { ApplicationOverview } from './ApplicationOverview'
 export { PaymentInfo } from './PaymentInfo'
 export { PaymentPending } from './PaymentPending'
-export { NationalIdWithName } from './NationalIdWithName'
+export { NationalIdWithName } from './NationalIdWithName/NationalIdWithName'
