+++ libs/application/templates/marriage-conditions/src/lib/messages.ts
@@ -486,4 +486,10 @@ export const m = defineMessages({
     defaultMessage: 'Lokadagsetning má ekki vera á undan upphafsdagsetningu',
     description: 'Error message if date is empty',
   },
+  nationalIdDuplicateError: {
+    id: 'mac.application:nationalIdDuplicateError',
+    defaultMessage:
+      'Kennitala hjá hjónaefni 2 má ekki vera sú sama og hjá hjónaefni 1',
+    description: '',
+  },
 })
